+++ src/YarnSpinner.Console/Commands/DumpTreeCommand.cs
@@ -12,8 +12,8 @@ namespace YarnSpinnerConsole
     {
         public static void DumpTree(FileInfo[] input, DirectoryInfo outputDirectory, bool json)
         {
-            input = CompileCommand.GetYarnFiles(input);
-            foreach (var inputFile in input)
+            var job = CompileCommand.GetCompilationJob(input);
+            foreach (var inputFile in job.Files.Select(file => new FileInfo(file.FileName)))
             {
                 var source = File.ReadAllText(inputFile.FullName);