+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - The metadata CSV file now stores its metadata in the same way as Yarn Spinner for Unity's export feature, with a single `metadata` column that contains a space-separated list of tags.
 - The `--allow-preview-features` (`-p`) flag can now be used when compiling scripts to enable preview features.
+- .ysls.json files referenced by Yarn Project files are now parsed and used when compiling Yarn code.
 
 ### Removed